move Fixture to test_fixture.rs
